mozreview: carry forward r+s for L3 request authors (Bug 1175166). r=mcote
If an author of a review request has level 3 commit access we trust that
they will re-request review when it is needed due to changes on an
already reviewed commit. We now formalize this by carrying forward any
r+s that a level 3 submitter had received before updating their review
request.