scripts: add a script to adjust repository permissions (bug 1186256); r=fubar
Inspection of repositories on production machines reveals incredible
inconsistency in filesystem permissions. Sanity injection is needed.

This script is roughly equivalent to:

  $ find repo -type f -exec chmod XXX {} \;
  $ find repo -type d -exec chmod XXXX {} \;

Although the file modes are configurable and it prints a summary of what
all changed. The summary is important: we can install this as a CRON and
it will only whine if permissions were adjusted. This will give insight
to server processes that are improperly adjusting permissions.