ansible/hg-web: install CRON to normalize repository permissions (bug 1186256); r=fubar
The repository permissions on the hgweb machines are all over the place.
While repositories are all likely read-write by the "hg" user, group
permissions are inconsistent. Permissions for others is also
inconsistent.

This commit creates a short shell script for iterating over repositories
in a base directory and adjusting permissions on them. We install a CRON
on the hgweb machines to perform this iteration daily.

We select permissions of 2775 for directories and 664 for files.