ansible/docker-hg-web: make hg-web a role dependency (bug 1187090); r=fubar
We want the docker-hg-web role to be built on top of hg-web, not for
there to be divergent roles. Make hg-web a role dependency of
docker-hg-web so we can start working towards this goal.

This does break the Docker hgweb environment. But some breakage is
needed to manage a change this large. The following commits will unbust
things.