ansible/hg-web: create /repo and /repo_local directories and symlinks (bug 1187090); r=fubar
We never added this to hg-web because the servers already had it. Now
that we're moving Docker to use the hg-web role, we need this logic in
hg-web.

Note that the docker-hg-web symlink was incorrect. Production has
/repo/hg -> /repo_local/mozilla. Also, Ansible's terminology around
links is also kinda weird. This is more confusing than it needs to be.