ansible/docker-hg-web: remove httpd configs (bug 1187090); r=fubar
hg-web already provides httpd config files and docker-hg-web's copies
are mostly that: redundant copies.

The httpd.conf file is identical with the exception of the ServerLimit and
MaxClients lines, which use templatized values in the case of hg-web. This
should "just work" on docker-hg-web. If not, we can fix it later.

The mod_wsgi.conf file content is identical between both versions.

vhost.conf does vary a bit. Manual inspection reveals this is mostly due
to hg-web's copy being updated and docker-hg-web's drifting out of sync.
Whatever the reason, we want to be based on hg-web, so remove the
docker-hg-web version.