docker: secure pulling of Docker base images (bug 1187544); r=smacleod
`docker pull` isn't secure because it doesn't verify digests *before*
processing incoming data. This opens the door for malicious images to
exploit a vulnerability in image application in the Docker daemon.
See https://titanous.com/posts/docker-insecurity for more.

In the near future, we are going to be using Docker to generate
images used to execute code on hg.mozilla.org. Importing malicious
Docker images (possibly as result of a MITM attack) could lead to subtle
compromise of the generation of these images, which could lead to
compromise of hg.mozilla.org. We introduce content based verification
of Docker base images to help prevent against this scenario.

We introduce a new special syntax to Dockerfiles. "FROM secure:"
defines a tuple of (repository, tag prefix, SHA-256, URL) defining
a remotely hosted image with known content. The "repository" and "tag
prefix" fields are used to locally tag the image so `docker images`
prints something sane.

When this syntax is encountered, we obtain the base image from the URL
specified, verify it, and import it. Or, we quickly obtain the image ID
by seeing if the hash of the (URL, SHA-256) already exists by scanning
local images. Compromise of the local Docker store could thus lead to
false cache hits. However, if your local Docker store is compromised,
your local system is already compromised and all bets are already off,
so we don't care.

We modify the code for "importing" the Docker image to use modified
Dockerfile content that references the ID of the imported image. Keep in
mind that Docker image IDs are not derived from just image content and
therefore not deterministic. That is why we can't hard-code the image
IDs in the Dockerfile. If we did, we could probably add special syntax
behind comments (e.g. "# sha256 XYZ" - like peep) and pass through the
Dockerfile unmodified. The unofficial pre-processed Dockerfile syntax is
unfortunate. But, it gets the job done.

We've modified relevant "FROM" statements in our in-tree Dockerfiles to
reference trusted Docker image archives. Because 3rd party Docker
registries and Git repositories containing these base images are
unreliable and may change from underneath us, I've uploaded these files
to S3 and inserted those URLs here.

The CentOS image comes from
https://github.com/CentOS/sig-cloud-instance-images at commit
311d80f2e558eba3a6ea88c387714ae2e4175702 and is simply a copy of
the docker/centos-6-20150615_2019-docker.tar.xz file in that revision.

The Ubuntu image is derived from
https://github.com/tianon/docker-brew-ubuntu-core at commit
4f7764b8eb70271b6a736341e090699fe058a198. Unlike the CentOS image,
modifications to the image need to occur. The image was built by running
`docker build .` in the trust/ directory of that repository to produce a
Docker image. When then created a container via `docker run -it <image>
/bin/false` so we could `docker export` that container to a tarball. The
xz compressed version of the tarball is what is uploaded. This process
may or may not be deterministic over time.