bzpost: properly handle unicode in user and description fields (bug 1187522); r=smacleod
This was more complicated than I thought it would be.

ctx.user() and ctx.description() return a special "local string" class
from mercurial.encoding. This behaves like a str but it is encoded in a
way that the locally configured encoding supports. So, unsupported code
points may render as "?" (or whatever the replacement character is). We
need to call encoding.fromlocal() to obtain the raw bytes (as a str).

Once this was done, I discovered that "baseuri" was a unicode type. This
means that the "lines" list consisted of mixed unicode and str and that
triggered an implicit conversion during the '\n'.join(), which failed
due to the non-ascii. This type is unicode because the module it comes
from is using unicode_literals. So I converted it to a str.