hgweb-chroot: create a C program to containerize moz.build evaluation (bug 1139218); r=kang
The hg.mozilla.org WSGI processes (running as the hg user) will need to
spawn a new, sandboxed process to evaluate moz.build files (read
potentially untrusted Python code). This commit introduces the executable
that starts the process.

Implemented is a C program that creates a specialized, ultra low
privileged container for moz.build evaluation. I found
http://crosbymichael.com/creating-containers-part-1.html and
http://www.unixwiz.net/techtips/chroot-practices.html invaluable when
implementing this.

The program is targeted for CentOS 6. CLONE_NEWUSER isn't exposed to our
servers, so we don't need to worry about that and the privilege
escalation bugs that come with it.