testing: establish a `d0cker` helper script
testing/docker-control.py is mach driver. All the other mach drivers
exist in the root directory. Having to type the full path to
testing/docker-control.py was annoying. Let's copy the script to `d0cker`
to make it easier to run.

We can't make the script named "docker" because it may conflict with the
official Docker CLI.