hgweb-chroot: generation of chroot environment (bug 1139218); r=smacleod
We need to evaluate some possibly untrusted Python code on the hgweb
servers. These servers are running CentOS 6. As much as I'd like to use
LXC, getting it to work on CentOS 6 is more or less an exercise in
futility. CentOS 6 is running Linux 2.6 and many of the nice features of
LXC (such as user namespaces) aren't available until later kernel
versions.

Until we run hg.mozilla.org on a modern OS and kernel, chroots seem to
be the easiest and most practical solution available to us.

This commit introduces a Docker image that is used to build files that
will be used for a chroot environment suitable for evaluating moz.build
files. We build Python 2.7.10 from source and install our minimum required
Python package dependencies. We manually copy shared libraries needed for
running Python. There are tools like jailkit that make creating
chroot environments very simple. However, the environment is currently
very simple, so it's not too difficult to manage, so I didn't bother
with the added complexity.

A mach command for obtaining files from the Docker image has been added.
We'll hook up these files to our deployment infrastructure in a
subsequent commit.