ansible/hg-web: configure moz.build chroot environment (bug 1139218); r=fubar, kang
We will be securing moz.build evaluation on hgweb machines with a chroot
and cgroups.

cgroups currently aren't configured on the hgweb machines. This commit
adds the Ansible bits to configure it.

We define a new control group - mozbuild - for evaluating mozbuild
files. We create a new "mozbuild" user and create a control group rule
that all processes for this user should be placed in the "mozbuild"
control group.

The "mozbuild" control group has some basic subsystem adjustments in
place to limit resource usage and device access.

Ansible has been updated to deploy our chroot files from the local
machine. It generates the current versions of the files and copies
them to the remote server.

We set up a bind mount of the repositories directory inside the chroot.
While the "mozbuild" user doesn't have permissions to modify anything
inside /repo/hg/mozilla (the entire tree is owned by hg:hg with
privileges 4775 on directories and 0664 on files), we add an extra layer
of caution by making the bind mount read-only to the chroot.

A sudoers policy to allow the "hg" user to invoke `mozbuild-eval` has
been added.