hgmo: support querying moz.build file info (bug 1139218); r=smacleod
The mozbuild Python package used by the Firefox build system recently
gained the ability to read moz.build files directly from Mercurial
repositories (bug 1168607). This opens the door to all kinds of exciting
possibilities for integrating moz.build info into Mercurial and tools
that are built on top of Mercurial, like code review (MozReview).

In this commit, we implement a generic Python module for extracting
moz.build file info from a Mercurial repository instance. We provide
a CLI command on the hgmo extension that prints obtained moz.build file
info. In future commits, we plan to call this CLI command from the
context of hgweb.

The output from the new command should be going through the Mercurial
templater. However, configuring this is extra work and there are no
users for non-JSON output yet. So, we defer this work until another
time.