hgmo: web command for retrieving moz.build info; r=smacleod
We provide a mechanism for exposing moz.build files info via the hgweb
server.

For security reasons that are explained in the added docs, we require a
wrapper command to look up moz.build files info. A subsequent commit
will implement a standalone command for establishing a secure
execution context for hg.mozilla.org servers.

We do not yet handle non-JSON output. Ideally, output should be going
through the templater so we can have HTML output of moz.build info.
But, hg.mozilla.org currently has a quirky relationship with JSON, as
the pushlog-legacy extension hard codes a JSON handler and interferes
with Mercurial's built-in templater. This will all need to be cleaned up
eventually.