ansible/hg-web: clean up package dependencies (bug 1189449); r=fubar
httpd should be explicitly listed, even though it seems to get pulled in
by httpd-devel.

We don't need the mod_wsgi package because we build mod_wsgi from source
and install it into the virtualenv. (Well, we don't install it from
source - Python's packaging installs it from source.)

The python-* packages aren't necessary, as we run things out of a
virtualenv now and the virtualenv has all the dependencies.