ansible/hg-web: move and refactor mirror-pull to hg-web (bug 1189449); r=fubar
The mirror-pull script is used to coordinate replication from master to
mirrors. We had a copy of the script already hooked up to the
docker-hg-web role, but it was done so very hackily. Basically we took
the .erb file from Puppet and checked it in as-is. We then modified the
file in Docker entrypoint script for that container. It was hacky.

We move mirror-pull to the hg-web role. Some of the variable
placeholders have been replaced with hard-coded values because they
don't vary in production nor in Docker. What template syntax remains we
convert to Jinja2. Default values for these variables have been added to
the hg-web role.

The entrypoint hackery in docker-hg-web still remains. It can be cleaned
up later.