ansible/hg-web: make all httpd prefork settings configurable (bug 1189449); r=fubar
It seems like a good idea to have knobs for all of these.

While I was here, I also removed the IfModule block around prefork
settings and the worker mpm settings completely. My thinking here is
that the server should *always* be in prefork mode and we want it to
fail fast if it is accidentally started with a different mpm.