ansible/hg-web: install WSGI files from version-control-tools (bug 1189449); r=fubar
The content of /repo/hg/webroot_wsgi was previously managed by hand.
There was little record keeping outside of bugs. And the process for
synchronizing changes across hosts was manual.

This commit does 2 things. First, we import the existing
/repo/hg/webroot_wsgi directory from its canonical home - hgssh1. The
contents of the users/ directory has been excluded from the import
because it is generated by a CRON job. Second, we add an Ansible task
for deploying these files. We use rsync with a "protect" rule to prevent
the contents of the users/ directory from being deleted.