ansible/hg: install mirror SSH key in hg-web role (bug 1189449); r=fubar
Freshly imaged machines need an SSH key installed for mirroring. While
we should ideally be fetching this from a secret store and using a
separate key per host, that's not how we do things today.

This commit establishes a hack of sorts to capture the SSH private key
from the hgssh1 host into a variable which is later written as part of
the hg-web role.