ansible/hg-web: create a virtualenv for tools (bug 1189449); r=fubar
We want to run more of our Python tooling with Python 2.7 and with
access to 3rd party Python libraries. This requires the use of a
virtualenv. Since the hgweb virtualenv is specialized and not
appropriate for general use, we create a new virtualenv to be used for
not hgweb things. Its use will become more apparent in subsequent
commits.