scripts: script to create a manifest of available repositories (bug 1189449); r=fubar
In order to efficiently bootstrap a new server, we want to clone several
repositories concurrently. In order to do this properly, we need to
respect certain repository properties, such as generaldelta.

We create a script that can produce a manifest of all known repositories
along with basic metadata. We'll use the output of this script to power
a tool for mass cloning/pulling all known repositories.