scripts: add script to efficiently sync hg repos from a manifest (bug 1189449); r=fubar
Currently, we have no effective way for mass cloning repositories to
machines. Previously, the most effective method we had was to rsync from
an existing server. However, rsync is not ideal from a Mercurial
perspective because it preserve the internal state of the repository,
including any ancient Mercurial settings ("requirements" in Mercurial
parlance) that were in effect when the repository was created for an
older Mercurial version. `hg clone`/`hg init` create new repositories
using the latest greatest feature set and `hg clone`/`hg pull` will auto
convert repository data to the newest/best version.

We establish a script that mass clones repositories. Give it a manifest
of repositories, a destination directory, and a base URL and it does the
rest. It will automatically use all available CPU cores for cloning. It
will preserve generaldelta on repositories.

The script integrates with the "mirror-pull" tool to synchronize hgrc
files. It should arguably be using mirror-pull for all cloning. However,
my intent is to eventually kill mirror-pull (at least in its current
form) because most of its functionality is not necessary, as "hg clone"
and/or "hg pull" or often sufficient. The script is overly complicated
and does things which aren't necessary, such as obtaining a replication
lock (Mercurial already has locking built in).