ansible/hg-web: remove hooks overrides when pulling (bug 1189938); r=fubar
All changegroup hooks were rewritten a few months ago so they no-op if
the current operation is a pull or strip (as opposed to a push). It is
therefore safe for all hooks to be active everywhere. So, the manual
disabling of these hooks in mirror-pull is no longer necessary.