ansible/hg-ssh: globally install push_printurls hook (bug 1189566); r=fubar
This hook is manually installed on many repos. It is generally useful.
Let's just install it globally so we don't have to continue spending
time managing it on all the repos.

The added config value matches exactly what exists in 94 repos on
hg.mozilla.org. Therefore, this is safe to deploy without any additional
changes to existing repository configs. There shouldn't be multiple
invocations of the hook or anything like that.