hghooks: deduce repository path from filesystem path (bug 1190494); r=RyanVM
Before, we maintained a manual mapping of repository directory name to
expanded path. This is a lot of extra work for something we can just
grab from the filesystem.

Refactor the hook to deduce the repository relative path from the
repository's location on the filesystem, relative to a root directory.

The test for an unknown repository not displaying info has been deleted
because the hook is now globally installed.