ansible/hg-web: move templates modification to hg-web role (bug 1190515); r=fubar
We want to pair down hgmo-extensions.yml into the common set of things
shared between hg-web and hg-ssh. Templates are hg-web only, so move
them to that role.

We change how the .vctnode file is referenced because when run from
Docker, it is no longer possible to reference a relative path. However,
pinning it from the {{ vct }} path seems to work.