ansible/hgmo: use hg-ssh role (bug 1190515); r=fubar
Up until now, all the hg-ssh work was only relevant to Docker because
we weren't using the hg-ssh role anywhere. With this commit, we finally
introduce the hg-ssh role into the hgmo deployment playbook.

The openssh-lpk role requires some LDAP variables. We grab these from
the remote JSON file. I'm not entirely confident this will actually
work, as the internet is a bit conflicted on the best way to do this.
The syntax for accessing JSON variables is quite hacky. Surely there is
a better way.