scripts: support reading list of repos from a file (bug 1191128); r=fubar
Instead of specifying multiple cron entries which may overlap if
execution is slow, it is safer to have a single process that handles all
repositories serially.

We add support to generate-hg-s3-bundles for reading the list of
repositories from a file. This will make it easy to define long lists of
repositories to generate bundles for.