ansible/hg-ssh: run S3 bundle generation out of virtualenv (bug 1191128); r=fubar
We now have a Python 2.7 virtualenv. Let's put it to use by having the
S3 bundle generation processes run out of it. This is a prerequisite to
introducing new features to the bundle generation script.

We remove the installation of python-boto from system packages and add
the boto package to the virtualenv. The transition should be
transparent.