pushlog: return an error if startID is too large (bug 1114843); r=edmorley
Clients need a mechanism to identify when a push ID value is out of
bounds. This patch gives them that.

This patch also introduces an error reporting mechanism for version 2
payloads. The new error mechanism may not be used everywhere yet. But it
should be. Failure to utilize the new error mechanism should be
considered a bug.