pushlog: move "info" web command to hgmo extension (bug 1171250); r=smacleod
hgwebjson.py has slowly been crippled. It is at the point where it only
defines the "info" web command. This command has nothing to do with
pushlog and has no business being in the pushlog-legacy directory. (This
directory held the contents of the old "pushlog" repository.)

We have the "hgmo" extension now, which is where we are placing
Mozilla-specific extensions for the hg.mozilla.org web server. It makes
the perfect host for the "info" web command.

Many files were changed because many things referred to this old
extension!