hgserver: add push tests
Now that we have a working hg.mozilla.org cluster in Docker, we can
finally start to write end-to-end tests encompassing hg.mozilla.org
behavior. \o/

We introduce a basic test for pushing and replication.

We create the "hgserver" top-level directory because we need a place to
hold tests. Eventually, I'd like to move other top-level directories
into this directory, as there are a number of directories related to
hg.mo and they should be consolidated.