hgserver: run hg from Python 2.7 virtualenv in pash (bug 1078442); r=fubar
Now that we have a Python 2.7 virtualenv, all we need to do is have
things use it.

We simply change pash to reference the `hg` in the virtualenv to have it
use the different Mercurial installation.

As part of this, we switch from Mercurial 3.4.1 to 3.4.2. Normally I
don't like doing multiple upgrades at once. But we already run 3.4.2 on
hgweb without problems and there is nothing in the changelog from 3.4.1
that warrants concern. We would have upgraded to 3.4.2 a long time ago
had an RPM been produced.