hgserver: add test verifying hooks work; r=fubar
We currently don't have any test coverage that hooks installed via
"python:mozhghooks" work as expected. Let's add some.

We don't need to verify the hooks do what they advertise: just that they
load and are active. Basically a smoke test.

As part of this, we add a support script to the Docker container for
easily adding hooks to a repository.