hgssh: reference mozhghooks from version-control-tools (bug 1196612); r=fubar
Python .pth files can be used to register additional search paths for
Python modules. Writing a .pth file is roughly functionally equivalent
to running `python setup.py develop`.

In this commit, we create a .pth file in the pash virtualenv that points
to the "hghooks" directory from the local version-control-tools
checkout. We also remove the addition of /repo/hg/libraries to the
Python search path of pash, effectively running everything from the
virtualenv.

We'll need to do something similar to hgweb to complete the transition.