ansible/hg-web: run hg from virtualenv in mirror-pull (bug 1196612); r=fubar
Now that we have a virtualenv tasked with handling replication, let's
use it.

This commit switches the mirror-pull script to invoke the hg executable
from the replication virtualenv. It also removes the PYTHONPATH
modification to register the /repo/hg/libraries directory because all
important content from that directory is available to the virtualenv.