hgserver: convert replication hook to a Python hook (bug 1196869); r=fubar
We want all hooks to be Python because, well, Python hooks are more
powerful.

As part of this, we have the hook print replication status, including
time taken to replicate. This will increase understanding of what takes
so long during pushes.

We change the names of the changegroup hooks to enforce sorting. And we
put URL printing after replication because it looks nicer.