mozreview: Override login page to use Bugzilla auth delegation (bug 993233). r=smacleod
Passing a "full" query-string parameter with any value, e.g. "full=true",
in the login URL results in the full login page being displayed. This exists
primarily to allow the admin user to log in, which does not have a
corresponding Bugzilla account.

Also override the "Log Out" link to not redirect to the login page, since if
you still have an open Bugzilla session, you'll be automatically logged back
into Review Board.

Selenium tests have been updated, but not all appear to be passing on tip
right now, so there are still failures.