autoland: fixup testing environment (bug 1198573) r=gps
This adds hglib to the test-requirements.txt file to support running the python
style unittest for Autoland. It fixes a crash on startup in autoland_pulse.py.
It also stops running autoland_pulse.py by default in the container and when
deploying to production as it currently does not do anything.