mozreview: display approval information in the commits table (Bug 1176321). r=dminor r=mdoglio
Now that we generate approval information for each review request we can
use this to show a status icon for each commit. This replaces the old
count of all ship-its on a review request which used to be displayed.

I chose to use r+ and r? to indicate approved / not approved since
Mozillians are familiar with that concept. While the semantics of
MozReview's approval don't match r+/? it should be an improvement
over the current status and we can iterate.