mozhg: support and prefer Bugzilla API keys (bug 1198087); r=smacleod
Bugzilla API keys are the preferred mechanism for defining credentials
because multiple keys can be generated and keys can (in theory) have
limited permissions.

Our first step towards using API keys is to have the mozhg auth
mechanism recognize them. We implement support for that in this commit.

We do not yet prompt for API keys if no credentials are defined. This
should be implemented once all downstream consumers support API keys.