reviewboard: modify test-auth.t to work in a API key world (bug 1198087); r=smacleod
As part of debugging test failures from later commits, it was realized
that the switch to API keys changed some assumptions in test-auth.t.
Before, we had to push to result in Review Board users being updated.
Now that `mozreview create-user` results in Bugzilla sending a ping back
to Review Board, this step isn't necessary. So, we remove some pushes.
We also fix a legacy password-based push that wasn't caught in the
previous commit.