mozreview: remove server support for passwords and cookies (bug 1198087); r=smacleod
API keys are the new and only supported way of specifying Bugzilla/Review
Board credentials. In this commit, we remove server-side support for
accepting passwords and cookies from clients.

If passwords or cookies are sent by the client, we fail fast and
instruct the client to change their client configuration.

We retain support in the client for sending passwords and cookies
because this commit will need to be deployed to production before we can
safely land removal of passwords and cookies from the client. Otherwise,
someone could upgrade to the password/cookie-less client before the
server knows how to speak API keys!