testing: run Mercurial commands in non-interactive mode; r=smacleod
If this is not specified, stdin from the test runner may get inherited
and commands that prompt may stall. This only impacts Selenium tests
since the .t test runner does special things with stdin. Surprisingly,
this was never an issue before.

Since publishing on push is the default behavior, the
test-reviewer-specification.py test no longer needs to publish. The code
for publishing and waiting on the changes after publishing has been removed.