mozreview: consolidate buttons for publishing reviews (bug 1195860); r=mdoglio
The review request page currently has many buttons/links related to
publishing/submitting review:

1) A "Review" text link in the actions list that opens the review
   dialog.
2) A "Ship It" text link in the actions list that submits the current
   review and marks it as ship it.
3) A "Edit Review" button in the review banner that opens the review
   dialog.
4) A "Publish" button in the review banner that publishes the review
   as-is, with no dialog for confirmation.

The redundancy in buttons and subtle nuances in behavior are confusing.
Many are confused what each button does. "Does it open a dialog for
additional confirmation or not?" "Why are there multiple review
buttons." "Reviews is a noun but it seems to perform an action."

This commit attempts to reduce confusion by:

1) Removing the "Publish" and "Ship It" buttons.
2) Changing the text of "Edit Review" to "Finish Review" so people have
   a hint that clicking it will finalize the review.
3) Changing the text of "Review" to "Finish Review" so it is a verb
   (like other text in the actions list) and so it matches the text
   in the review banner.

The new UI is simpler and consistent.

The simplicity does come at a price: power users no longer can click a
single button to publish or ship it a review. Instead, they must open
the review dialog and click "Publish Review" there. Since many new users
were confused by the various buttons, I think this trade-off is
acceptable. If people miss the old functionality, we can always add it
back in somehow, preferably with better UI to indicate what exactly is
going to happen. For now, we favor simplicity.