mozreview: Don't return summaries of discarded review requests (bug 1200881). r=smacleod
Bug 1200615 documents an error in mozreview.extra_data.get_parent_rr() in
which the wrong parent is returned for a child review request with a
discarded parent.  The proper fix is moving the extra_data structure into a
proper table; for now, we just omit discarded review requests, which aren't
really useful to summarize anyway.