pash: make LDAP access date updating actually work (bug 1200847); r=fubar
Previously, if the "hgAccessDate" attribute was missing from the LDAP
entry, we would never create it.

The new logic looks to see if the entry has the "hgAccount" objectClass.
If it does, it can hold a "hgAccessDate" attribute and we attempt to
define it.

For good measure, tests verifying the addition of the hgAccessDate
attribute have been added.