ansible/hg-reviewboard: install hgserver's version of sh_helper.py
Currently, hg.mozilla.org and reviewboard-hg.mozilla.org share similar
but different versions of our Mercurial SSH shell "pash." This commit
marks the beginning of an effort to consolidate to a single pash
implementation by removing the hg-reviewboard instance of sh_helper.py.
This file is completely identical in behavior. The only source changes
are around hgserver's support for adding periods in prompts.