ansible/hg-reviewboard: run hgweb from a virtualenv
We do this on hg.mozilla.org. The virtualenv is better isolated and runs
from Python 2.7.

As part of this, some other cleanup was performed. Basically bringing
hg-reviewboard inline with hgweb. We should eventually create a role for
the HTTP server pieces of a Mercurial host so there isn't so much
duplication.