vcsreplicator: create project
This commit introduces the "vcsreplicator" project / Python package. The
goal of this project is to create a robust version control replication
service that serves the immediate business needs of Mozilla (and
possibly others).

This commit is more than a skeleton. It contains basic code for
obtaining a Kafka producer suitable for recording events of interest. It
also includes the beginnings of a Mercurial extension to write events
into Kafka.

A basic test showing that the config subsystem works has been included.