vcsreplicator: add a mechanism to map repo paths to pull URLs
It is not guaranteed for producers to know the URL that consumers should
pull data from. This is why the repository URL is not recorded in the
message indicating there is changegroup data.

This commit introduces a mapping table that allows consumers to rewrite
repository paths to URLs for the purposes of constructing a URL to be
used for pulling.