hgmo: web command for info about repo divergence (bug 1199459); r=smacleod
We want MozReview to display info about how a review has "diverged" from
upstream. Initially, we're going to track number of commits behind and
files that have potential conflicts.

This commit implements a web command that accepts a base revision and
set of paths and returns number of commits to the latest head and any
files that may potentially be in conflict.