mozreview: fix diff display when hostingservice is added to a repo (bug 1208213) r=mcote
Diff display was broken due to the truncation of the base_commit_id argument
to get_file in the HostingService base class so that we would always retrieve
the tip revision, causing patching to fail. This overrides get_file and
get_file_exists to pass in *kwargs to avoid truncation of arguments.