mozreview: make commit syntax for specifying reviewers case-insensitive (bug 1201853) r=gps
We have a few users who have capitals in their ircnicks. Bugzilla (and the
MozReview UI) will auto-complete these properly, but using r? in the syntax
fails to do so leading to confusion.

This adds a loop over the results of get_users (which is case insensitive)
first looking for an exact match, and if that fails, a case insensitive match.
The call to get_user is removed, as it is always case sensitive.