mozreview: always display try button but disable it with tooltip when appropriate (Bug 1198915) r=mdoglio r=dminor
Originally we would hide the entire automation menu when a user wasn't
allowed to trigger a try build. This is problematic for adding the
autoland button to the menu because it has different rules for when
to show.

Now instead of hiding the whole menu, we disable all entries by default
and provide the reason in a tooltip. This is more informative and allows
us to add more entries with different access rules.