mozreview: add a script that validates Bugzilla API Keys (bug 1195856); r=smacleod, kang
We're going to configure the existing MozReview Mercurial server to
allow pushes over HTTP. We'll be using standard HTTP Basic
authentication using Bugzilla usernames and API Keys. In order to
authenticate these credentials, we're going to use mod_authnz_external
to invoke a script that determines whether the credentials are valid.
This is that script.