testing: allowing testing with local reviewboard checkout (bug 1174805) r=gps
This is a partial solution to allowing testing with a local reviewboard
checkout. It adds a mach command that will replace the version of reviewboard
running on the rbweb instance with one built from a local path.

This is mostly useful for deploying a hacked up version of reviewboard which
matches the version of reviewboard already in use on the container to
facillitate debugging reviewboard code. The full solution would be to add
support for building the container using reviewboard from a local checkout in
the first place.