reviewboard: preserve Unicode in commit messages (bug 1213884); r=dminor
There were 2 bugs preventing Unicode in commit messages from working.

First, ctx.description() returns a mercurial.encoding.localstr instance.
This is a thin wrapper around str that quacks like a str but has the
value represented in the local encoding, which is variable and may not
be able to represent non-ASCII characters, at which point they are
represented as ASCII question marks.

Second, we were feeding this localstr or a str instance into RBTools.
RBTools appears to assume values are unicode because it blindly
performs a .encode('utf-8') on them. This operation should be illegal
(you .encode() unicode and .decode() str/bytes) but Python 2 happily
allows us.

Combined with a string concatenation, the bowels of RBTools encountered
a UnicodeDecodeError when Unicode was present in commit messages.

With this change, Unicode is now properly preserved in the Review Board
summary and description fields as well as in Bugzilla attachment names,
as the test changes demonstrate.