mozreview: support pushing via HTTP using Bugzilla credentials (bug 1195856); r=fubar, smacleod
Currently, only people with an LDAP account can push to MozReview (over
SSH). The barrier to obtaining an LDAP account is quite high and is not
suitable for first-time contributors.

This commit enables pushing to MozReview via HTTP using Bugzilla
usernames and API keys as credentials.

The Apache httpd server has been configured to authenticate non {GET,
HEAD} requests using the previously written Bugzilla API Key authentication
script using mod_authnz_external.

We had to modify the hgrc on the hgrb server to both accept pushes over
HTTP and not require TLS for those pushes.

Because the hgweb server now needs to talk to Review Board, we had to
add various Python packages to the hgweb virtualenv.

Because the LDAP and Bugzilla user databases are separate, we added
prefixing of HTTP/Bugzilla users in the pushlog so we can distinguish
which database each user comes from. We retain unprefixed LDAP entries
for backwards compatibility. No machines should be looking at the
pushlog on MozReview, so we could probably rewrite the pushlogs at some
later date to contain prefixes. Or we could just add them now and deal
with both formats for LDAP pushers until we rewrite...

Comprehensive tests of the functionality have been added.

There are a few minor areas where HTTP pushing is sub-par. They will be
addressed in subsequent commits.