reviewboard: automatically use Bugzilla credentials from hgrc; r=smacleod, kang
We already have users define Bugzilla usernames and API Keys in their
hgrc file under the "bugzilla.username" and "bugzilla.apikey" options.

Mercurial has its own method for storing credentials for URLs. To make
initial setup easier and so Bugzilla API Keys aren't duplicated in the
config files, we add integration into Mercurial's password manager that
will automatically use [bugzilla] defined credentials when talking to a
known Mozilla HTTP server.

We establish a list of trusted URL endpoints that we release the
credentials to. Unless an attacker convinces someone to change the
trusted list and connect to a rogue server (it's hard to prevent
phishing), the current implementation can only be exploited by
someone conducting a MitM, with a x509 certificate for a mozilla.org
hostname, and with that certificate signed by a CA trusted by the
client. This should be very unlikely. Keep in mind that if someone
is in a position to carry out this attack, they are probably in position
to intercept Bugzilla and other Mozilla traffic as well, so the Bugzilla
credentials can already be considered compromised. This is a long winded
way of saying I believe the current implementation provides sufficient
security.