reviewboard: refactor repository discovery to facilitate HTTP pushing; r=smacleod
When we implemented review repository discovery, we assumed there would
only ever be a single URL we'd push to - the SSH URL. Now that we allow
pushes over HTTP, there are multiple URLs for each review repo.

We refactor how the server exposes review repos so it advertises both an
HTTP and SSH URL for each repo.

Since we had to create a new client capability so old clients wouldn't
choke on the new format, I also took the time to change how the data is
exposed to clients. Before, we were using the "pushkey" protocol. I'm no
longer a fan of shoving everything inside the pushkey protocol, so I
established a dedicated wire protocol command for exposing the raw file
contents. I dropped the old capability because the listkeys command to
retrieve data from it will return an empty set. It shouldn't be called
by modern clients since modern clients will require an extension upgrade
before we get to this code (due to the introduced client-side
capability).

A test for discovery over HTTP has been added. It discovered that
capabilities were being cached across repo redirects. This has also been
fixed.